generator/requirements.txt
pdf2image
pytesseract
torch
transformers
Pillow